+++ lily-design-system-angular-date-time-picker/examples/basic.component.ts
@@ -52,6 +52,10 @@ export class BasicExample {
   readonly labels: DateTimePickerLabels = {
     previousYear: "Previous year",
     previousMonth: "Previous month",
+    previousWeek: "Previous week",
+    previousDay: "Previous day",
+    nextDay: "Next day",
+    nextWeek: "Next week",
     nextMonth: "Next month",
     nextYear: "Next year",
     confirm: "OK",